[level 0] Title: 다음에 올 숫자, Time: 0.02 ms, Memory: 75.7 MB -BaekjoonHub

+++ 프로그래머스/lv0/120924. 다음에 올 숫자/다음에 올 숫자.java
@@ -3,6 +3,6 @@ class Solution {
         if (common[1] - common[0] == common[2] - common[1])
             return common[common.length - 1] + common[1] - common[0];
         
-        return (int)((double)common[common.length - 1] * common[1] / common[0]);
+        return common[common.length - 1] * common[1] / common[0];
     }
 }
